workforceaustralia/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
webdriver-manager==4.0.1
selenium==4.22.0
cloudscraper==1.2.71
pydantic==2.11.5
sqlalchemy==2.0.41  
cryptography==45.0.2
beautifulsoup4==4.13.4
gspread==6.2.1
google-auth==2.40.2
openai==1.84.0